Reword intro to Api Overview to flow better
Not sure if this is too dramatic. This is an attempt to address Issue #34, i.e. that "document set" was a bit unnatural. 

Decided to give a short sharp answer in the beginning, and tack on that initial sentence to the end of this section, which felt more natural as the preceding sentences explained the why behind it nicely.