Made intro section more generic and concise
The intented results is a little vague. I think the intended result is for the user to experience your product working, so I tried to word it more that way.

I think the second point now captures what I deleted regarding making the benefits of your product apparent to the user (not just telling them, but showing them).

Attempts to address Issue #37