Resolves [Issue 39](https://github.com/thegooddocsproject/templates/issues/39)
Changed references to "discussion" doc type to "explanation" doc type, primarily in the explanation directory (formerly the discussion directory). Also some minor incidental wordsmithing along the way.

Search for references to "discussion" or "discussions" in the templates directory. There are now very few, and the ones that remain are still appropriate despite the renaming of this doc type.